Ampache: With APIv5, the "singular" entities are not wrapped in JSON array
While using the JSON APIv5+, the responses of "singular" actions (like
`song`, `artist`) are different from "plural" actions (like `songs`,
`artists`). For singular actions, the responses are like:
{ id: "1", name: "Something", ... }
For plural actions (even with just one result), the responses are like:
{ song: [ { id: "1", name: "Something", ... }, ... ] }

With JSON APIv4, the responses of both singular and plural actions are
like:
[ { id: "1", name: "Something", ... }, ... ]

In the XML API, there are no such differences.