Ampache: Use cache-friendly `art` URLs on response objects
The `art` URLs returned within entity objects no longer contain the
sessions `auth` token. Instead, we generate own token for each image
which depends on the entity type, entity id, and the API key. This way,
the URLs don't change between sessions but revoking the API key still
invalidates all old URLs.

This new cache-friendly URL looks like
https://<cloud_url>/index.php/apps/music/ampache/image.php?object_type=album&object_id=<id>&token=<token>
This is similar to the URL used by the real Ampache server which is
important to be able to support
[ample](https://github.com/mitchray/ample) client which assumes such
URL to exist and give a placeholder when called with dummy args.