Ampache: Terminate related sessions when API key revoked
As the previous commit gave us the relation between Ampache sessions
and and API keys (stored in the table music_ampache_users), it was now
trivial to improve the safety of the API key revocation: All ongoing
sessions related to the revoked API key are now immediately terminated.
Previously, the existing sessions could continue to function normally
as long as the client kept them alive.